修复钉钉配置页未勾选自动获取token时，多次保存配置会报错问题

+++ ali_dindin/models/res_config_settings.py
@@ -82,5 +82,6 @@ class ResConfigSettings(models.TransientModel):
         else:
             cron = self.env['ir.cron'].sudo().search(
                 [('code', '=', "env['ali.dindin.get.token'].get_token()")])
-            cron.sudo().unlink()
+            if len(cron) >= 1:
+                cron.sudo().unlink()
